# Static Foreign Calls