--- a/unison-runtime/src/Unison/Runtime/ANF.hs
+++ b/unison-runtime/src/Unison/Runtime/ANF.hs
@@ -42,7 +42,7 @@ module Unison.Runtime.ANF
     SuperGroup (..),
     arities,
     POp (..),
-    FOp,
+    ForeignFunc(..),
     close,
     saturate,
     float,
@@ -1030,12 +1030,12 @@ pattern TPrm ::
   ABTN.Term ANormalF v
 pattern TPrm p args = TApp (FPrim (Left p)) args
 
-pattern AFOp :: FOp -> [v] -> ANormalF v e
+pattern AFOp :: ForeignFunc -> [v] -> ANormalF v e
 pattern AFOp p args = AApp (FPrim (Right p)) args
 
 pattern TFOp ::
   (ABT.Var v) =>
-  FOp ->
+  ForeignFunc ->
   [v] ->
   ABTN.Term ANormalF v
 pattern TFOp p args = TApp (FPrim (Right p)) args
@@ -1232,9 +1232,6 @@ instance Semigroup (BranchAccum v) where
 instance Monoid (BranchAccum e) where
   mempty = AccumEmpty
 
--- Foreign operation, indexed by words
-type FOp = Word64
-
 data Func v
   = -- variable
     FVar v
@@ -1247,7 +1244,7 @@ data Func v
   | -- ability request
     FReq !Reference !CTag
   | -- prim op
-    FPrim (Either POp FOp)
+    FPrim (Either POp ForeignFunc)
   deriving (Show, Eq, Functor, Foldable, Traversable)
 
 data Lit
@@ -1437,6 +1434,256 @@ data POp
   | NOTB -- not
   | ANDB -- and
   | IORB -- or
+  deriving (Show, Eq, Ord, Enum, Bounded)
+
+-- | Enum representing every foreign call.
+data ForeignFunc
+  = IO_UDP_clientSocket_impl_v1
+  | IO_UDP_UDPSocket_recv_impl_v1
+  | IO_UDP_UDPSocket_send_impl_v1
+  | IO_UDP_UDPSocket_close_impl_v1
+  | IO_UDP_ListenSocket_close_impl_v1
+  | IO_UDP_UDPSocket_toText_impl_v1
+  | IO_UDP_serverSocket_impl_v1
+  | IO_UDP_ListenSocket_toText_impl_v1
+  | IO_UDP_ListenSocket_recvFrom_impl_v1
+  | IO_UDP_ClientSockAddr_toText_v1
+  | IO_UDP_ListenSocket_sendTo_impl_v1
+  | IO_openFile_impl_v3
+  | IO_closeFile_impl_v3
+  | IO_isFileEOF_impl_v3
+  | IO_isFileOpen_impl_v3
+  | IO_getEcho_impl_v1
+  | IO_ready_impl_v1
+  | IO_getChar_impl_v1
+  | IO_isSeekable_impl_v3
+  | IO_seekHandle_impl_v3
+  | IO_handlePosition_impl_v3
+  | IO_getBuffering_impl_v3
+  | IO_setBuffering_impl_v3
+  | IO_setEcho_impl_v1
+  | IO_getLine_impl_v1
+  | IO_getBytes_impl_v3
+  | IO_getSomeBytes_impl_v1
+  | IO_putBytes_impl_v3
+  | IO_systemTime_impl_v3
+  | IO_systemTimeMicroseconds_v1
+  | Clock_internals_monotonic_v1
+  | Clock_internals_realtime_v1
+  | Clock_internals_processCPUTime_v1
+  | Clock_internals_threadCPUTime_v1
+  | Clock_internals_sec_v1
+  | Clock_internals_nsec_v1
+  | Clock_internals_systemTimeZone_v1
+  | IO_getTempDirectory_impl_v3
+  | IO_createTempDirectory_impl_v3
+  | IO_getCurrentDirectory_impl_v3
+  | IO_setCurrentDirectory_impl_v3
+  | IO_fileExists_impl_v3
+  | IO_getEnv_impl_v1
+  | IO_getArgs_impl_v1
+  | IO_isDirectory_impl_v3
+  | IO_createDirectory_impl_v3
+  | IO_removeDirectory_impl_v3
+  | IO_renameDirectory_impl_v3
+  | IO_directoryContents_impl_v3
+  | IO_removeFile_impl_v3
+  | IO_renameFile_impl_v3
+  | IO_getFileTimestamp_impl_v3
+  | IO_getFileSize_impl_v3
+  | IO_serverSocket_impl_v3
+  | Socket_toText
+  | Handle_toText
+  | ThreadId_toText
+  | IO_socketPort_impl_v3
+  | IO_listen_impl_v3
+  | IO_clientSocket_impl_v3
+  | IO_closeSocket_impl_v3
+  | IO_socketAccept_impl_v3
+  | IO_socketSend_impl_v3
+  | IO_socketReceive_impl_v3
+  | IO_kill_impl_v3
+  | IO_delay_impl_v3
+  | IO_stdHandle
+  | IO_process_call
+  | IO_process_start
+  | IO_process_kill
+  | IO_process_wait
+  | IO_process_exitCode
+  | MVar_new
+  | MVar_newEmpty_v2
+  | MVar_take_impl_v3
+  | MVar_tryTake
+  | MVar_put_impl_v3
+  | MVar_tryPut_impl_v3
+  | MVar_swap_impl_v3
+  | MVar_isEmpty
+  | MVar_read_impl_v3
+  | MVar_tryRead_impl_v3
+  | Char_toText
+  | Text_repeat
+  | Text_reverse
+  | Text_toUppercase
+  | Text_toLowercase
+  | Text_toUtf8
+  | Text_fromUtf8_impl_v3
+  | Tls_ClientConfig_default
+  | Tls_ServerConfig_default
+  | Tls_ClientConfig_certificates_set
+  | Tls_ServerConfig_certificates_set
+  | TVar_new
+  | TVar_read
+  | TVar_write
+  | TVar_newIO
+  | TVar_readIO
+  | TVar_swap
+  | STM_retry
+  | Promise_new
+  | Promise_read
+  | Promise_tryRead
+  | Promise_write
+  | Tls_newClient_impl_v3
+  | Tls_newServer_impl_v3
+  | Tls_handshake_impl_v3
+  | Tls_send_impl_v3
+  | Tls_decodeCert_impl_v3
+  | Tls_encodeCert
+  | Tls_decodePrivateKey
+  | Tls_encodePrivateKey
+  | Tls_receive_impl_v3
+  | Tls_terminate_impl_v3
+  | Code_validateLinks
+  | Code_dependencies
+  | Code_serialize
+  | Code_deserialize
+  | Code_display
+  | Value_dependencies
+  | Value_serialize
+  | Value_deserialize
+  | Crypto_HashAlgorithm_Sha3_512
+  | Crypto_HashAlgorithm_Sha3_256
+  | Crypto_HashAlgorithm_Sha2_512
+  | Crypto_HashAlgorithm_Sha2_256
+  | Crypto_HashAlgorithm_Sha1
+  | Crypto_HashAlgorithm_Blake2b_512
+  | Crypto_HashAlgorithm_Blake2b_256
+  | Crypto_HashAlgorithm_Blake2s_256
+  | Crypto_HashAlgorithm_Md5
+  | Crypto_hashBytes
+  | Crypto_hmacBytes
+  | Crypto_hash
+  | Crypto_hmac
+  | Crypto_Ed25519_sign_impl
+  | Crypto_Ed25519_verify_impl
+  | Crypto_Rsa_sign_impl
+  | Crypto_Rsa_verify_impl
+  | Universal_murmurHash
+  | IO_randomBytes
+  | Bytes_zlib_compress
+  | Bytes_gzip_compress
+  | Bytes_zlib_decompress
+  | Bytes_gzip_decompress
+  | Bytes_toBase16
+  | Bytes_toBase32
+  | Bytes_toBase64
+  | Bytes_toBase64UrlUnpadded
+  | Bytes_fromBase16
+  | Bytes_fromBase32
+  | Bytes_fromBase64
+  | Bytes_fromBase64UrlUnpadded
+  | Bytes_decodeNat64be
+  | Bytes_decodeNat64le
+  | Bytes_decodeNat32be
+  | Bytes_decodeNat32le
+  | Bytes_decodeNat16be
+  | Bytes_decodeNat16le
+  | Bytes_encodeNat64be
+  | Bytes_encodeNat64le
+  | Bytes_encodeNat32be
+  | Bytes_encodeNat32le
+  | Bytes_encodeNat16be
+  | Bytes_encodeNat16le
+  | MutableArray_copyTo_force
+  | MutableByteArray_copyTo_force
+  | ImmutableArray_copyTo_force
+  | ImmutableArray_size
+  | MutableArray_size
+  | ImmutableByteArray_size
+  | MutableByteArray_size
+  | ImmutableByteArray_copyTo_force
+  | MutableArray_read
+  | MutableByteArray_read8
+  | MutableByteArray_read16be
+  | MutableByteArray_read24be
+  | MutableByteArray_read32be
+  | MutableByteArray_read40be
+  | MutableByteArray_read64be
+  | MutableArray_write
+  | MutableByteArray_write8
+  | MutableByteArray_write16be
+  | MutableByteArray_write32be
+  | MutableByteArray_write64be
+  | ImmutableArray_read
+  | ImmutableByteArray_read8
+  | ImmutableByteArray_read16be
+  | ImmutableByteArray_read24be
+  | ImmutableByteArray_read32be
+  | ImmutableByteArray_read40be
+  | ImmutableByteArray_read64be
+  | MutableByteArray_freeze_force
+  | MutableArray_freeze_force
+  | MutableByteArray_freeze
+  | MutableArray_freeze
+  | MutableByteArray_length
+  | ImmutableByteArray_length
+  | IO_array
+  | IO_arrayOf
+  | IO_bytearray
+  | IO_bytearrayOf
+  | Scope_array
+  | Scope_arrayOf
+  | Scope_bytearray
+  | Scope_bytearrayOf
+  | Text_patterns_literal
+  | Text_patterns_digit
+  | Text_patterns_letter
+  | Text_patterns_space
+  | Text_patterns_punctuation
+  | Text_patterns_anyChar
+  | Text_patterns_eof
+  | Text_patterns_charRange
+  | Text_patterns_notCharRange
+  | Text_patterns_charIn
+  | Text_patterns_notCharIn
+  | Pattern_many
+  | Pattern_many_corrected
+  | Pattern_capture
+  | Pattern_captureAs
+  | Pattern_join
+  | Pattern_or
+  | Pattern_replicate
+  | Pattern_run
+  | Pattern_isMatch
+  | Char_Class_any
+  | Char_Class_not
+  | Char_Class_and
+  | Char_Class_or
+  | Char_Class_range
+  | Char_Class_anyOf
+  | Char_Class_alphanumeric
+  | Char_Class_upper
+  | Char_Class_lower
+  | Char_Class_whitespace
+  | Char_Class_control
+  | Char_Class_printable
+  | Char_Class_mark
+  | Char_Class_number
+  | Char_Class_punctuation
+  | Char_Class_symbol
+  | Char_Class_separator
+  | Char_Class_letter
+  | Char_Class_is
+  | Text_patterns_char
   deriving (Show, Eq, Ord, Enum, Bounded)
 
 type ANormal = ABTN.Term ANormalF