--- a/unison-runtime/src/Unison/Runtime/ANF/Serialize.hs
+++ b/unison-runtime/src/Unison/Runtime/ANF/Serialize.hs
@@ -26,7 +26,6 @@ import Unison.Reference (Reference, Reference' (Builtin), pattern Derived)
 import Unison.Runtime.ANF as ANF hiding (Tag)
 import Unison.Runtime.Exception
 import Unison.Runtime.Serialize
-import Unison.Util.EnumContainers qualified as EC
 import Unison.Util.Text qualified as Util.Text
 import Unison.Var (Type (ANFBlank), Var (..))
 import Prelude hiding (getChar, putChar)
@@ -317,7 +316,7 @@ putGroup ::
   (MonadPut m) =>
   (Var v) =>
   Map Reference Word64 ->
-  EC.EnumMap ForeignFunc Text ->
+  Map ForeignFunc Text ->
   SuperGroup v ->
   m ()
 putGroup refrep fops (Rec bs e) =
@@ -338,7 +337,7 @@ getGroup = do
   cs <- replicateM l (getComb ctx n)
   Rec (zip vs cs) <$> getComb ctx n
 
-putCode :: (MonadPut m) => EC.EnumMap ForeignFunc Text -> Code -> m ()
+putCode :: (MonadPut m) => Map ForeignFunc Text -> Code -> m ()
 putCode fops (CodeRep g c) = putGroup mempty fops g *> putCacheability c
 
 getCode :: (MonadGet m) => Word32 -> m Code
@@ -363,7 +362,7 @@ putComb ::
   (MonadPut m) =>
   (Var v) =>
   Map Reference Word64 ->
-  EC.EnumMap ForeignFunc Text ->
+  Map ForeignFunc Text ->
   [v] ->
   SuperNormal v ->
   m ()
@@ -384,7 +383,7 @@ putNormal ::
   (MonadPut m) =>
   (Var v) =>
   Map Reference Word64 ->
-  EC.EnumMap ForeignFunc Text ->
+  Map ForeignFunc Text ->
   [v] ->
   ANormal v ->
   m ()
@@ -482,7 +481,7 @@ putFunc ::
   (MonadPut m) =>
   (Var v) =>
   Map Reference Word64 ->
-  EC.EnumMap ForeignFunc Text ->
+  Map ForeignFunc Text ->
   [v] ->
   Func v ->
   m ()
@@ -496,7 +495,7 @@ putFunc refrep fops ctx f = case f of
   FReq r c -> putTag FReqT *> putReference r *> putCTag c
   FPrim (Left p) -> putTag FPrimT *> putPOp p
   FPrim (Right f)
-    | Just nm <- EC.lookup f fops ->
+    | Just nm <- Map.lookup f fops ->
         putTag FForeignT *> putText nm
     | otherwise ->
         exn $ "putFunc: could not serialize foreign operation: " ++ show f
@@ -757,7 +756,7 @@ putBranches ::
   (MonadPut m) =>
   (Var v) =>
   Map Reference Word64 ->
-  EC.EnumMap ForeignFunc Text ->
+  Map ForeignFunc Text ->
   [v] ->
   Branched (ANormal v) ->
   m ()
@@ -825,7 +824,7 @@ putCase ::
   (MonadPut m) =>
   (Var v) =>
   Map Reference Word64 ->
-  EC.EnumMap ForeignFunc Text ->
+  Map ForeignFunc Text ->
   [v] ->
   ([Mem], ANormal v) ->
   m ()
@@ -997,7 +996,7 @@ deserializeCode bs = runGetS (getVersion >>= getCode) bs
         n | 1 <= n && n <= 3 -> pure n
         n -> fail $ "deserializeGroup: unknown version: " ++ show n
 
-serializeCode :: EC.EnumMap ForeignFunc Text -> Code -> ByteString
+serializeCode :: Map ForeignFunc Text -> Code -> ByteString
 serializeCode fops co = runPutS (putVersion *> putCode fops co)
   where
     putVersion = putWord32be codeVersion
@@ -1023,7 +1022,7 @@ serializeCode fops co = runPutS (putVersion *> putCode fops co)
 -- shouldn't be subject to rehashing.
 serializeGroupForRehash ::
   (Var v) =>
-  EC.EnumMap ForeignFunc Text ->
+  Map ForeignFunc Text ->
   Reference ->
   SuperGroup v ->
   L.ByteString