--- a/unison-runtime/unison-runtime.cabal
+++ b/unison-runtime/unison-runtime.cabal
@@ -44,6 +44,7 @@ library
       Unison.Runtime.Exception
       Unison.Runtime.Foreign
       Unison.Runtime.Foreign.Function
+      Unison.Runtime.Foreign.Impl
       Unison.Runtime.Interface
       Unison.Runtime.IOSource
       Unison.Runtime.Machine