## Overview

Foreign Calls were being dispatched dynamically. This puts all foreign calls in a big'ol case statement and inlines all the ForeignConvention typeclasses.

Interestingly, The first iteration of this PR was _slower_ than trunk, but explicitly NOINLINE'ing the foreign calls (with a wrapper to ensure Stack gets unboxed) sped it up significantly. This implies that code size of `eval` was affecting things code-caching significantly, which isn't too surprising considering these are all microbenchmarks. So we should be careful about how much code we have in our main loop; maybe even should consider splitting off some of the lesser used instructions into their own chunk of code.

It's actually quite surprising that this change now _speeds up_ the regular suite, despite it not using foreign calls for most of its benches.

## Implementation notes

* Make a big ol' case statement for every foreign call
* Use that instead of FF records
* Adds a new Sandbox Failure instruction
* Move sandboxing for foreigns to a pre-processing step over all source code which removes foreign calls and replaces them with an instruction that just errors.
* Fix sandboxing of Ref instructions which were added previously


## Test coverage

* [x] Tested that the new foreign call sandboxing works as expected.

Benchmarks

```
foreigns : '{IO, Exception} ()
foreigns = do
  printTime
    "time-foreigns" 10 (n ->
      repeat n do
        repeat 1000 do
          use Either toBug
          t = toBug <| monotonic.impl()
          _ = sec t
          _ = nsec t
          -- repeat ~100 times
  tvar = TVar.newIO 0
  printTime
    "tvar foreigns" 10 (n ->
      repeat n do
        repeat 1000 do
          atomically do
            _ = read tvar
            _ = write tvar 1
            _ = swap tvar 10
            ~repeat ~100 times

```

trunk -> new

```
Time foreigns
25.45208ms -> 23.89372ms

tvar foreigns
42.9391ms -> 36.974ms
```

Normal suite, I wouldn't have expected these to change much since they don't use foreigns very much, but it looks like things still somehow improved a bit 🎉 

```
fib1
325.198µs -> 314.807µs

fib2
2.293117ms -> 2.259432ms

fib3
2.671696ms -> 2.646835ms

Decode Nat
345ns -> 337ns

Generate 100 random numbers
210.18µs -> 207.108µs

List.foldLeft
2.141224ms -> 2.025939ms

Count to 1 million
128.4097ms -> 124.7779ms

Json parsing (per document)
268.28µs -> 261.794µs

Count to N (per element)
191ns -> 188ns

Count to 1000
192.286µs -> 188.466µs

Mutate a Ref 1000 times
318.636µs -> 314.031µs

CAS an IO.ref 1000 times
427.475µs -> 425.158µs

List.range (per element)
332ns -> 325ns

List.range 0 1000
349.66µs -> 345.243µs

Set.fromList (range 0 1000)
1.681155ms -> 1.592714ms

Map.fromList (range 0 1000)
1.242585ms -> 1.173887ms

NatMap.fromList (range 0 1000)
4.986759ms -> 4.871296ms

Map.lookup (1k element map)
2.617µs -> 2.546µs

Map.insert (1k element map)
7.114µs -> 6.888µs

List.at (1k element list)
285ns -> 277ns

Text.split /
32.893µs -> 32.688µs
```

## Loose ends

Probably worth investigating the overhead of the sandbox checks in the interpreter, we may be able to remove them for instructions as well.